refactor: add new utility functions for credential preparation and metadata sanitization
